LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/fetch-request-css-cross-origin.https-expected.txt

CONSOLE MESSAGE: Did not parse stylesheet at 'https://web-platform.test:9443/service-workers/service-worker/resources/cross-origin-html.css?mime=no' because non CSS MIME types are not allowed for cross-origin stylesheets.

PASS setup global state
PASS MIME checking of CSS resources fetched via service worker when Content-Type is not set.
FAIL Same-origin policy for access to CSS resources fetched via service worker promise_test: Unhandled rejection with value: object "SecurityError: Not allowed to access cross-origin stylesheet"
PASS cleanup global state
